Add option to clean build; have it only build fv3, soca, iodaconv

+++ build.sh
@@ -19,6 +19,7 @@ usage() {
   echo "  -t  target to build for <target>    DEFAULT: $(hostname)"
   echo "  -c  additional CMake options        DEFAULT: <none>"
   echo "  -v  build with verbose output       DEFAULT: NO"
+  echo "  -f  force a clean build             DEFAULT: NO"
   echo "  -d  include JCSDA ctest data        DEFAULT: NO"
   echo "  -h  display this message and quit"
   echo
@@ -33,6 +34,7 @@ CMAKE_OPTS=""
 BUILD_TARGET="$(hostname)"
 BUILD_VERBOSE="NO"
 CLONE_JCSDADATA="NO"
+CLEAN_BUILD="NO"
 
 while getopts "p:t:c:hvd" opt; do
   case $opt in
@@ -50,6 +52,9 @@ while getopts "p:t:c:hvd" opt; do
       ;;
     d)
       CLONE_JCSDADATA=YES
+      ;;
+    f)
+      CLEAN_BUILD=YES
       ;;
     h|\?|:)
       usage
@@ -79,7 +84,9 @@ case ${BUILD_TARGET} in
 esac
 
 BUILD_DIR=${BUILD_DIR:-$dir_root/build}
-[[ -d ${BUILD_DIR} ]] && rm -rf ${BUILD_DIR}
+if [[ $CLEAN_BUILD == 'YES' ]]; then
+  [[ -d ${BUILD_DIR} ]] && rm -rf ${BUILD_DIR}
+fi
 mkdir -p ${BUILD_DIR} && cd ${BUILD_DIR}
 
 # If INSTALL_PREFIX is not empty; install at INSTALL_PREFIX
@@ -95,8 +102,13 @@ set +x
 
 # Build
 echo "Building ..."
+builddirs="fv3-jedi soca iodaconv"
 set -x
-make -j ${BUILD_JOBS:-6} VERBOSE=$BUILD_VERBOSE
+for b in $builddirs; do
+  cd $b
+  make -j ${BUILD_JOBS:-6} VERBOSE=$BUILD_VERBOSE
+  cd ../
+done
 set +x
 
 # Install
